requirements.txt
boto3
flask
flatten_json
pymongo
pytest
pyyaml
